requirements.txt
pandas==1.2.4
numpy==1.21.6
Pillow==7.2.0
transformers==4.25.1
tensorflow==2.9.2
torch==1.9.0
opencv-python==4.5.1.48
matplotlib==3.3.2
Flask==2.1.0
flask-restx==1.0.3
